requirements.txt
Flask==3.0.3
gunicorn==20.1.0
requests==2.26.0
apscheduler==3.8.1
scikit-image==0.24.0
Pillow==11.1.0
tzlocal==2.1
